pyproject.toml
[project]
name = "tata-chemicals-v2v-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.14"
dependencies = [
    "alembic",
    "asyncpg",
    "beautifulsoup4",
    "celery",
    "fastapi",
    "greenlet>=3.3.0",
    "livekit>=1.0",
    "livekit-agents[aws,cartesia,deepgram,openai,sarvam,silero,turn-detector]~=1.3",
    "pydantic-settings",
    "pwdlib[argon2]",
    "pyjwt",
    "python-dotenv~=1.0",
    "python-multipart",
    "redis",
    "requests",
    "sqlalchemy",
    "uvicorn",
]
